+++ spec/models/draw_spec.rb
@@ -35,6 +35,12 @@ RSpec.describe Draw, type: :model do
       draw.destroy
       expect(student.reload.old_draw_id).to be_nil
     end
+    it 'clears groups on destruction' do
+      group = FactoryGirl.create(:group)
+      draw = FactoryGirl.create(:draw, group_ids: group.id)
+      draw.destroy
+      expect { group.reload }.to raise_error(ActiveRecord::RecordNotFound)
+    end
   end
 
   describe '#suite_sizes' do